requirements.txt
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.37.1
botocore==1.37.1
branca==0.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==37.0.2
fastapi==0.115.11
firebase-admin==6.7.0
frozenlist==1.5.0
GeoAlchemy2==0.17.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.3.2
numpy==2.2.4
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pika==1.3.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
s3transfer==0.11.3
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
xyzservices==2025.1.0
yarl==1.18.3
